LPS-55800 Move resources
