+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-taglib/src/META-INF/liferay-ddm.xml
