--- util-taglib/src/com/liferay/taglib/ddm/base/packageinfo
@@ -1,1 +1,0 @@
-version 7.0.0
