--- util-taglib/src/com/liferay/taglib/ddm/packageinfo
@@ -1,1 +1,0 @@
-version 7.0.0
